app/ingest/pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "ptinsight-ingest"
version = "1.0"
description = "PT Insight Ingest"
authors = ["Dominik Stiller <dominik.stiller@hpe.com>"]
packages = [
    { include = "ptinsight" },
]

[tool.poetry.dependencies]
python = "^3.8"
paho-mqtt = "^1.5.0"
kafka-python = "^2.0.1"
pyyaml = "^5.3.1"
boto3 = "^1.14.20"
ptinsight-common = {path = "../common/python", develop=false}

[tool.poetry.dev-dependencies]
black = ">=19.10b0"
